Issue #151 Adding an extra cast to prevent S2I conversion saves 17% off binsrch. Strictly speaking the result won't be the same, but for this benchmark it will be. Doing the same cast in the C version doesn't change anything to the C performance.

+++ src/lib/bm_binsrch/c/common/javax_rtcbench_RTCBenchmark.c
@@ -22,7 +22,7 @@ void __attribute__((noinline)) rtcbenchmark_measure_native_performance(uint16_t 
         uint16_t high = NUMNUMBERS - 1;
         uint16_t mid;
         while (low <= high) {
-            mid = (low + high) >> 1;
+            mid = ((uint16_t)(low + high)) >> 1;
             int16_t number_mid;
             if ((number_mid=numbers[mid]) < toFind) {
                 low = mid + 1;
